Series of bug fixes related to the hunter drone.
Bug fix - chain detonation (HunterDroneBehaviour.cs): DetectExternalCollision now skips colliders whose parent has a HunterDroneBehaviour. Drones from the same player no longer trigger each other.

Dead code removed (HunterDroneBehaviour.cs): HandleHit collapsed from 14 lines to 4 - the HitsRequired <= 0, HitCount >= HitsRequired, and HitCount++ checks were all unreachable. HitsRequired = 1; HitCount = 0; removed from Start since HandleHit no longer uses them.

_shotDown removed (HunterDroneBehaviour.cs): Field and all references deleted. _exploded alone is sufficient - both HandleHit and ShootDown guard on it.

Field placement fixed (HunterDroneBehaviour.cs): MaxFlightDistance and MaxSpeed moved into the Configuration section with the other public fields. Internal state section now contains only private state.

Re-validation moved to FixedUpdate (HunterDroneBehaviour.cs): Re-validation now runs before the retry timer block, so the flow reads top-to-bottom: re-validate → maybe retry → move. Removed from UpdateHoming.

Guard order fixed in ServerLaunchDrone (HunterDroneNetworkBridge.cs): Item/prefab validation runs first; the concurrent-drone cap check comes after, with LogInfo instead of LogWarning.

Minor redundancies removed (HunterDroneBehaviour.cs): _collisionBuffer[i] == null check, ThrowerInfo != null guard in IsValidTarget, and the dead magnitude check in DirectionToRotation all removed.

Uses typed correctly (HunterDroneConfig.cs, HunterDroneItemDefinition.cs): ConfigEntry<float> → ConfigEntry<int>, default 1f → 1, cast removed from MaxUses.

+++ Config/HunterDroneConfig.cs
@@ -8,7 +8,7 @@ namespace IssaPlugin
         private const string Section = "HunterDrone";
 
         public ConfigEntry<Key> GiveKey { get; private set; }
-        public ConfigEntry<float> Uses { get; private set; }
+        public ConfigEntry<int> Uses { get; private set; }
         public ConfigEntry<float> LaunchSpeed { get; private set; }
         public ConfigEntry<float> Acceleration { get; private set; }
         public ConfigEntry<float> HomingStopDistance { get; private set; }
@@ -29,7 +29,7 @@ namespace IssaPlugin
                 Key.NumpadPlus,
                 "Debug key to add the Hunter Drone item to your inventory."
             );
-            Uses = cfg.Bind(Section, "Uses", 1f, "Number of uses per Hunter Drone pickup.");
+            Uses = cfg.Bind(Section, "Uses", 1, "Number of uses per Hunter Drone pickup.");
             LaunchSpeed = cfg.Bind(
                 Section,
                 "LaunchSpeed",
